requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==20.8b1
click==7.1.2
Django==3.2
django-restframework==0.0.1
mypy-extensions==0.4.3
pathspec==0.8.1
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
